+++ src/Order/AddToCartCommandInterface.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Setono\SyliusGiftCardPlugin\Order;
 
 use Setono\SyliusGiftCardPlugin\Model\OrderInterface;
-use Setono\SyliusGiftCardPlugin\Order\Dto\AddGiftCardToCartInformationInterface;
 use Sylius\Bundle\OrderBundle\Controller\AddToCartCommandInterface as BaseCommandInterface;
 use Sylius\Component\Core\Model\OrderItemInterface;
 
@@ -15,5 +14,5 @@ interface AddToCartCommandInterface extends BaseCommandInterface
 
     public function getCartItem(): OrderItemInterface;
 
-    public function getGiftCardInformation(): AddGiftCardToCartInformationInterface;
+    public function getGiftCardInformation(): GiftCardInformationInterface;
 }
