--- src/Order/Dto/Factory/AddGiftCardToCartInformationFactoryInterface.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Setono\SyliusGiftCardPlugin\Order\Dto\Factory;
-
-use Setono\SyliusGiftCardPlugin\Order\Dto\AddGiftCardToCartInformationInterface;
-use Sylius\Component\Order\Model\OrderItemInterface;
-
-interface AddGiftCardToCartInformationFactoryInterface
-{
-    public function createNew(OrderItemInterface $orderItem): AddGiftCardToCartInformationInterface;
-}
