+++ src/Order/Factory/AddToCartCommandFactory.php
@@ -5,20 +5,20 @@ declare(strict_types=1);
 namespace Setono\SyliusGiftCardPlugin\Order\Factory;
 
 use Setono\SyliusGiftCardPlugin\Order\AddToCartCommandInterface;
-use Setono\SyliusGiftCardPlugin\Order\Dto\Factory\AddGiftCardToCartInformationFactoryInterface;
 use Sylius\Bundle\OrderBundle\Factory\AddToCartCommandFactoryInterface;
 use Sylius\Component\Order\Model\OrderInterface;
 use Sylius\Component\Order\Model\OrderItemInterface;
 
 final class AddToCartCommandFactory implements AddToCartCommandFactoryInterface
 {
+    /** @var class-string<AddToCartCommandInterface> */
     private string $className;
 
-    private AddGiftCardToCartInformationFactoryInterface $extraInformationFactory;
+    private GiftCardInformationFactoryInterface $extraInformationFactory;
 
     public function __construct(
         string $className,
-        AddGiftCardToCartInformationFactoryInterface $extraInformationFactory
+        GiftCardInformationFactoryInterface $extraInformationFactory
     ) {
         $this->className = $className;
         $this->extraInformationFactory = $extraInformationFactory;
@@ -26,12 +26,6 @@ final class AddToCartCommandFactory implements AddToCartCommandFactoryInterface
 
     public function createWithCartAndCartItem(OrderInterface $cart, OrderItemInterface $cartItem): AddToCartCommandInterface
     {
-        /**
-         * @var AddToCartCommandInterface $addToCartCommand
-         * @psalm-suppress InvalidStringClass
-         */
-        $addToCartCommand = new $this->className($cart, $cartItem, $this->extraInformationFactory->createNew($cartItem));
-
-        return $addToCartCommand;
+        return new $this->className($cart, $cartItem, $this->extraInformationFactory->createNew($cartItem));
     }
 }
