+++ src/Resources/config/services/order.xml
@@ -3,22 +3,22 @@
            xmlns="http://symfony.com/schema/dic/services"
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <parameters>
-        <parameter key="setono_sylius_gift_card.order.model.add_gift_card_to_cart_information.class">Setono\SyliusGiftCardPlugin\Order\Dto\AddGiftCardToCartInformation</parameter>
+        <parameter key="setono_sylius_gift_card.order.model.gift_card_information.class">Setono\SyliusGiftCardPlugin\Order\GiftCardInformation</parameter>
         <parameter key="setono_sylius_gift_card.order.model.add_to_cart_command.class">Setono\SyliusGiftCardPlugin\Order\AddToCartCommand</parameter>
     </parameters>
 
     <services>
-        <service id="setono_sylius_gift_card.order.factory.add_gift_card_to_cart_information"
-                 class="Setono\SyliusGiftCardPlugin\Order\Dto\Factory\AddGiftCardToCartInformationFactory">
-            <argument>%setono_sylius_gift_card.order.model.add_gift_card_to_cart_information.class%</argument>
+        <service id="setono_sylius_gift_card.order.factory.gift_card_information"
+                 class="Setono\SyliusGiftCardPlugin\Order\Factory\GiftCardInformationFactory">
+            <argument>%setono_sylius_gift_card.order.model.gift_card_information.class%</argument>
         </service>
-        <service id="Setono\SyliusGiftCardPlugin\Order\Dto\Factory\AddGiftCardToCartInformationFactoryInterface"
-                 alias="setono_sylius_gift_card.order.factory.add_gift_card_to_cart_information"/>
+        <service id="Setono\SyliusGiftCardPlugin\Order\Factory\GiftCardInformationFactoryInterface"
+                 alias="setono_sylius_gift_card.order.factory.gift_card_information"/>
 
         <service id="sylius.factory.add_to_cart_command" class="Setono\SyliusGiftCardPlugin\Order\Factory\AddToCartCommandFactory"
                  public="true">
             <argument>%setono_sylius_gift_card.order.model.add_to_cart_command.class%</argument>
-            <argument type="service" id="setono_sylius_gift_card.order.factory.add_gift_card_to_cart_information" />
+            <argument type="service" id="setono_sylius_gift_card.order.factory.gift_card_information" />
         </service>
 
         <service id="sylius.form.type.add_to_cart" class="Sylius\Bundle\CoreBundle\Form\Type\Order\AddToCartType">
